modify according to MR.Wang
